chore: Set bundled NODE_ENV default to production.
